qcacmn: Fix compilation issue on arch32 by gcc 9.3
This change fixed compilation error about implicit-fallthrough and
pointer to in cast.

CRs-Fixed: 2814658

+++ hal/wifi3.0/hal_srng.c
@@ -220,7 +220,7 @@ QDF_STATUS hal_construct_shadow_regs(void *hal_soc)
 			shadow_config_index;
 		hal->list_shadow_reg_config[i].va =
 			SHADOW_REGISTER(shadow_config_index) +
-			(uint64_t)hal->dev_base_addr;
+			(uintptr_t)hal->dev_base_addr;
 		hal_debug("target_reg %x, shadow register 0x%x shadow_index 0x%x",
 			  hal->shadow_config[shadow_config_index].addr,
 			  SHADOW_REGISTER(shadow_config_index),
